backend/crc-api/api-container/requirements.txt
fastapi
uvicorn
azure-data-tables